pyproject.toml
[project]
name = "pixel-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scipy>=1.15.2",
    "streamlit>=1.43.0",
]
